add support for open api spec 2.11.0
