+++ package.json
@@ -1,5 +1,5 @@
 {
   "dependencies": {
-    "@okta/openapi": "2.10.0"
+    "@okta/openapi": "2.11.0"
   }
 }
